MICCAI24/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
absl-py=2.0.0=pyhd8ed1ab_0
aiohttp=3.9.0=pypi_0
aiosignal=1.3.1=pyhd8ed1ab_0
alabaster=0.7.16=pypi_0
annotated-types=0.6.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.0=pyhd8ed1ab_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
autopep8=2.0.4=pypi_0
babel=2.14.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.1=openblas
blinker=1.7.0=pyhd8ed1ab_0
brotli=1.0.9=pypi_0
brotli-python=1.0.9=py310hd8f1fbe_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.11.17=hbcca054_0
cachetools=5.3.2=pyhd8ed1ab_0
certifi=2023.7.22=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.3.0=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.1.1=pypi_0
cryptography=41.0.3=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.2.140=0
cuda-runtime=12.1.0=0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
dgl=1.1.2=pypi_0
dglgo=0.0.2=pypi_0
docker-pycreds=0.4.0=pypi_0
docutils=0.20.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.12.4=pypi_0
fonttools=4.43.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=pypi_0
fsspec=2023.9.2=pypi_0
future=0.18.3=pyhd8ed1ab_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.40=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.26.1=pyhca7485f_0
google-auth-oauthlib=1.2.0=pyhd8ed1ab_0
grpc-cpp=1.48.2=h5bf31a4_0
grpcio=1.48.2=pypi_0
idna=3.4=pyhd8ed1ab_0
imagesize=1.4.1=pypi_0
importlib-metadata=7.0.1=pyha770c72_0
ipykernel=6.14.0=pypi_0
ipython=8.4.0=pypi_0
isort=5.13.2=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h5eee18b_1
jupyter-core=5.4.0=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.4.0=py310hff52083_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=15_linux64_openblas
libcblas=3.9.0=15_linux64_openblas
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
liblapack=3.9.0=15_linux64_openblas
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libopenblas=0.3.20=pthreads_h78a6416_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
littleutils=0.2.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.1=pyhd8ed1ab_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.0=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=pypi_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.24.4=pypi_0
numpydoc=1.6.0=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-ml-py3=7.352.0=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.2.140=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pyhd8ed1ab_0
ogb=1.3.6=pypi_0
openblas=0.3.20=pthreads_h320a7e8_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
outdated=0.2.2=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.0.1=pypi_0
pip=23.3=pypi_0
platformdirs=3.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
protobuf=4.24.4=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.1=pyhd8ed1ab_0
pyasn1-modules=0.3.0=pyhd8ed1ab_0
pycodestyle=2.11.1=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pyg=2.4.0=py310_torch_1.12.0_cu113
pygcl=0.1.2=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_0
pyopenssl=23.2.0=pyhd8ed1ab_1
pyparsing=3.1.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.12.1=cpu_py310h9dbd814_1
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
pyzmq=25.1.1=pypi_0
rdkit-pypi=2022.9.5=pypi_0
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
ruamel-yaml=0.18.5=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.1=pypi_0
seaborn=0.13.1=pypi_0
sentry-sdk=1.32.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.0.0=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
snowballstemmer=2.2.0=pypi_0
sphinx=7.2.6=pypi_0
sphinxcontrib-applehelp=1.0.7=pypi_0
sphinxcontrib-devhelp=1.0.5=pypi_0
sphinxcontrib-htmlhelp=2.0.4=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.6=pypi_0
sphinxcontrib-serializinghtml=1.1.9=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.15.1=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=pypi_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=2.1.0=pypi_0
torch-cluster=1.6.3+pt21cu121=pypi_0
torch-geometric=2.4.0=pypi_0
torch-scatter=2.1.2+pt21cu121=pypi_0
torch-sparse=0.6.18+pt21cu121=pypi_0
torchaudio=2.1.0=pypi_0
torchvision=0.16.0=pypi_0
tornado=6.1=pypi_0
tqdm=4.66.1=pyhd8ed1ab_0
traitlets=5.11.2=pyhd8ed1ab_0
triton=2.1.0=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023.3=pypi_0
urllib3=2.0.7=pypi_0
wandb=0.15.12=pypi_0
wcwidth=0.2.8=pyhd8ed1ab_0
werkzeug=3.0.1=pyhd8ed1ab_0
wheel=0.41.2=pypi_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
yarl=1.7.2=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zmq=0.0.0=pypi_0
zstd=1.5.5=hc292b87_0
